feature: allow song overlay
