+++ lib/stream_merger.rb
@@ -7,6 +7,7 @@ require "securerandom"
 require "tempfile"
 require "time"
 require_relative "stream_merger/concat"
+require_relative "stream_merger/merger"
 require_relative "stream_merger/merger_utils"
 require_relative "stream_merger/stream_file"
 require_relative "stream_merger/utils"
